enhancement/13 Add docs and comments
